Organize merge/imessage, merge/path -> read
