+++ src/sql/asset.sql
