+++ src/sql/text.sql
